Add files via upload
When dealing with data having two parameters varying it is necessary to eliminate at least one. I have written a code that helps obtain the radius and coordinates of a large number of particles having different radii. Generally studying large-size data with 1000 particles and having 3000 plus frames is difficult. In comes the necessity to automate the work. I used a type of thresholding called binary thresholding and morphological opening to clean my image and make the image easier for computer analysis . Using OpenCV to make images better for processing is essential and the key takeaway is that you have to play around with parameters to suit your image and choose the right type of thresholding also inverting the image helped the detection process. This code saves the effort that would be required to train a neural network . All you need to do is modify it accordingly for the parameter you want to relate. I used coordinates which were a linker between the parameter and the radius . You need to upload the images that you wish to analyse.The code will create files of its contours,thresholds and Excel file of its data .Then you need to feed in data having the  coordinates related to the parameter and let the code run and the end with a zip file you will have an excel folder having the mapped data.